Update color_game_of_life.m
I commented out "% imagesc(current_gen); %creates an initial generation matrix of random 0’s and 1’s" and replaced this with a new variable I called "colorswitch" which creates a 3D matrix that randomizes the 'current_gen' while taking in the 'colorswitch' matrix I created. This took me 3 hours to research and teach myself about, and then another 1 1/2 hours to implement and debug.