Update color_check_neighbors.m
I updated this code to include 3 cycles of randomization with 'new_gen' to keep the game running longer, and to prevent the colors from quickly interacting then stopping. This implementation took 1 hour to research, another 1 hour to implement, and 2 hours of debugging and ultimately discussing my issues/bugs with my group in Gathertown for an hour this morning.